refactor(cloud): move some static util methods outside `CloudApi` class
