+++ core/src/commands/login.ts
@@ -14,7 +14,7 @@ import type { AuthTokenResponse } from "../cloud/api.js"
 import { CloudApi, getGardenCloudDomain } from "../cloud/api.js"
 import type { Log } from "../logger/log-entry.js"
 import { ConfigurationError, TimeoutError, InternalError, CloudApiError } from "../exceptions.js"
-import { AuthRedirectServer } from "../cloud/auth.js"
+import { AuthRedirectServer, saveAuthToken } from "../cloud/auth.js"
 import type { EventBus } from "../events/events.js"
 import type { ProjectConfig } from "../config/project.js"
 import { findProjectConfig } from "../config/base.js"
@@ -103,7 +103,7 @@ export class LoginCommand extends Command<{}, Opts> {
 
     log.info({ msg: `Logging in to ${cloudDomain}...` })
     const tokenResponse = await login(log, cloudDomain, garden.events)
-    await CloudApi.saveAuthToken(log, globalConfigStore, tokenResponse, cloudDomain)
+    await saveAuthToken(log, globalConfigStore, tokenResponse, cloudDomain)
     log.success({ msg: `Successfully logged in to ${cloudDomain}.`, showDuration: false })
 
     return {}
