+++ core/src/commands/logout.ts
@@ -16,6 +16,7 @@ import { ConfigurationError } from "../exceptions.js"
 import type { ProjectConfig } from "../config/project.js"
 import { findProjectConfig } from "../config/base.js"
 import { BooleanParameter } from "../cli/params.js"
+import { clearAuthToken } from "../cloud/auth.js"
 
 export const logoutOpts = {
   "disable-project-check": new BooleanParameter({
@@ -92,7 +93,7 @@ export class LogOutCommand extends Command<{}, Opts> {
       `
       log.warn(msg)
     } finally {
-      await CloudApi.clearAuthToken(log, garden.globalConfigStore, cloudDomain)
+      await clearAuthToken(log, garden.globalConfigStore, cloudDomain)
       log.success(`Successfully logged out from ${cloudDomain}.`)
     }
     return {}
